+++ src/baseq2rtxp/sharedgame/sg_time.h
@@ -61,7 +61,12 @@ public:
 	}
 
 	int64_t frames( ) const {
+		#ifdef CLGAME_INCLUDE
+		return _ms / clgi.frame_time_ms;
+		#endif
+		#ifdef SVGAME_INCLUDE
 		return _ms / gi.frame_time_ms;
+		#endif
 	}
 
 	// check if non-zero